# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Library Management API",
    "description": "API for managing a library's books collection",
    "version": "1.0.0"
  },
  "host": "localhost:5000",
  "basePath": "/",
  "schemes": ["http"],
  "paths": {
    "/books": {
      "get": {
        "summary": "List all books",
        "description": "Retrieve a list of all books in the library.",
        "responses": {
          "200": {
            "description": "List of books",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Book"
              }
            }
          }
        }
      },
      "post": {
        "summary": "Add a new book",
        "description": "Add a book to the library.",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Book"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Book added successfully"
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/books/{isbn}": {
      "get": {
        "summary": "Get book by ISBN",
        "description": "Retrieve a specific book by ISBN.",
        "parameters": [
          {
            "name": "isbn",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Book details",
            "schema": {
              "$ref": "#/definitions/Book"
            }
          },
          "404": {
            "description": "Book not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a book by ISBN",
        "description": "Remove a book from the library by its ISBN.",
        "parameters": [
          {
            "name": "isbn",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Book deleted successfully"
          },
          "404": {
            "description": "Book not found"
          }
        }
      },
      "put": {
        "summary": "Update book details",
        "description": "Update the details of a book by ISBN.",
        "parameters": [
          {
            "name": "isbn",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Book"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Book updated successfully"
          },
          "404": {
            "description": "Book not found"
          }
        }
      }
    }
  },
  "definitions": {
    "Book": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string"
        },
        "author": {
          "type": "string"
        },
        "published_year": {
          "type": "integer"
        },
        "isbn": {
          "type": "string"
        },
        "genre": {
          "type": "string"
        }
      },
      "required": ["title", "author", "published_year", "isbn"]
    }
  }
}